requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
coverage==6.5.0
crispy-bootstrap5==0.7
dj-database-url==1.0.0
Django==4.1.3
django-axes==5.40.1
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-coverage==1.2.4
django-crispy-forms==1.14.0
django-environ==0.9.0
django-factory-boy==1.0.0
django-filter==22.1
django-ipware==4.0.2
django-js-asset==2.0.0
django-lockout==0.1.1
django-nose==1.4.7
django-on-heroku==1.1.2
django-rest-framework==0.1.0
django-richtextfield==1.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
fake-factory==9999.9.9
Faker==15.3.3
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
nose==1.3.7
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
whitenoise==6.2.0
